chatapp_backend/requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt
mysqlclient==2.2.7
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.13.0
gunicorn
whitenoise
django-allauth
urllib3
pycparser
oauthlib
idna
defusedxml
charset-normalizer
certifi
requests
python3-openid
cffi
requests-oauthlib
cryptography
social-auth-core
social-auth-app-django
dj-rest-auth[social]
django-cors-headers
channels==4.0.0
channels_redis==4.2.0
daphne==4.1.0